+++ src/vector3.py
@@ -4,9 +4,9 @@ import math
 
 @dataclass
 class Vector3:
-    x: int = 0
-    y: int = 0
-    z: int = 0
+    x: float = 0
+    y: float = 0
+    z: float = 0
 
     def __str__(self) -> str:
         return f"Vector3({self.x}, {self.y}, {self.z})"
